Now using an EventBus to pass timeout and alarm events to the SurveyActivity. Depending on the state of the app, onNewIntent() was sometimes getting called after onResume(), and so events weren’t consistently handled.
Also added an AlarmBroadcastReceiver that is useful for testing out alarms. This can be used like so from the command line:

adb shell am broadcast -a com.askonthego.alarm -e cron "52 * * * MON"

In order to run the alarm immediately, run the following command:

adb shell am broadcast -a com.askonthego.alarm -e cron "52 * * * MON" --ez runNow true

Note that the cron expression is ignored with this version of the command.